+++ lib/src/auth/components/custom_text_form_field.dart
@@ -1,5 +1,3 @@
-import 'package:flutter/src/foundation/key.dart';
-import 'package:flutter/src/widgets/framework.dart';
 import 'package:flutter/material.dart';
 
 class CustomFormTextField extends StatefulWidget {
