--- lib/src/models/page_manager.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-class PageManager {
-  PageManager(this._pageController);
-
-  final PageController _pageController;
-
-  int page = 0;
-
-
-  void setPage(int value) {
-    if(value == page) return;
-    page = value;
-    _pageController.jumpToPage(value);
-  }
-}
