--- lib/src/screen/login_screen.dart
@@ -1,37 +1,0 @@
-import 'package:flutter/src/foundation/key.dart';
-import 'package:flutter/src/widgets/framework.dart';
-import 'package:flutter/material.dart';
-import 'package:lojavirtual/src/auth/components/custom_form_Sign_in_screen.dart';
-import 'package:lojavirtual/src/auth/components/custom_text_logo.dart';
-
-class LoginScreen extends StatefulWidget {
-  const LoginScreen({Key? key}) : super(key: key);
-
-  @override
-  State<LoginScreen> createState() => _LoginScreenState();
-}
-
-class _LoginScreenState extends State<LoginScreen> {
-  @override
-  Widget build(BuildContext context) {
-    final size = MediaQuery.of(context).size;
-    
-    return Scaffold(
-      backgroundColor: const Color(0xFF7D79D0),
-      body: SingleChildScrollView(
-        child: SizedBox(
-          height: size.height,
-          width: size.width,
-          child: Column(
-            children: <Widget>[
-              // Logo
-              const CustomLogo(),
-              // Formulário
-              FormSignInScreen(),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
