+++ lib/src/screen/screen.dart
@@ -1,9 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/src/foundation/key.dart';
-import 'package:flutter/src/widgets/framework.dart';
-import 'package:lojavirtual/src/models/page_manager.dart';
-import 'package:lojavirtual/src/screen/login_screen.dart';
-import 'package:provider/provider.dart';
 
 class Screen extends StatelessWidget {
   final PageController pageController = PageController();
